Add script for collapsing tracked ARs to single samples
Runs through the different artmip datasets and for each unique AR,
selects the relevant time steps and sums the time variable. Creates a
sample id that contains the ar_id and the start and end time. Also
saves info about all the samples (across ardts) in a single csv file.